coconots: Convolution-Closed Models for Count Time Series

Useful tools for fitting, validating, and forecasting of practical convolution-closed time series models for low counts are provided. Marginal distributions of the data can be modelled via Poisson and Generalized Poisson innovations. Regression effects can be incorporated through time varying innovation rates. The models are described in Jung and Tremayne (2011) doi:10.1111/j.1467-9892.2010.00697.x and the model assessment tools are presented in Czado et al. (2009) doi:10.1111/j.1541-0420.2009.01191.x and, Tsay (1992) doi:10.2307/2347612 .

Author

Maintainer: Manuel Huth manuel.huth@yahoo.com

Authors:

Robert C. Jung

Andy Tremayne